New methods for object handler.
The QObjectHandler class now allows for methods with no return value in
which case the method is responsible for writing directly to the socket.
Also the QObjectHandler POST and PUT methods can be called without the
explicit body argument, in which case the reading should be made within
the method.