Don't close the socket when payload is read manually
Avoiding closing the socket when the payload is to be read manually
because it could be a big payload and continuously computation may need
to take place and the main loop needs to be involved.

Reading and writting from/to the socket will take advantage of the
signals and slots.

Note: the socket needs to be close explicitly when the computation ends.